% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateConversion.R
\docType{data}
\name{data}
\alias{data}
\alias{DegreeConv}
\alias{value}
\title{This function takes a matrix or data frame that contains a set of latitudes and longitudes
in degree notation. It will convert the degree coordinates to decimal degrees and reassign
the given matrix to the newly converted values. There is also an option for direction.
When set to true, the function will give the user a direction
calculcated by the average of all points.}
\format{
An object of class \code{data.frame} with 12001 rows and 5 columns.
}
\usage{
DegreeConv(long_data, lat_data, Direction = FALSE)
}
\arguments{
\item{long_data}{a matrix or data frame}

\item{lat_data}{a matrix or data frame}

\item{Direction}{enables users to add a mean direction if curious}

\item{value}{A list or data frame of converted data}
}
\value{
matrix thats contains converted decimal degree values
}
\description{
This function takes a matrix or data frame that contains a set of latitudes and longitudes
in degree notation. It will convert the degree coordinates to decimal degrees and reassign
the given matrix to the newly converted values. There is also an option for direction.
When set to true, the function will give the user a direction
calculcated by the average of all points.
}
\keyword{datasets}
